relicense under MIT license
commit logs shows commits from mboehn (me), kradalby and laaknor.

laaknor have two commits, which is one merge and one adding a name
to the README file. I am not a lawyer, but these commit does not
seem to be copyrighted works as of Norwegian copyright law.

Both mboehn (myself) and kradalby agrees to relicense to the MIT
license. kradalby confirmed this today in a project related Slack
channel.